+++ data-raw/bde_ind_db.R
@@ -18,7 +18,6 @@ bde_ind_db_init <- tribble(
   mutate(Numero_secuencial = as.character(Numero_secuencial)) |>
   arrange(tidyBdE_fun)
 
-
 # Add metadata
 full_cat <- bde_catalog_load()
 bde_cat <- full_cat |>