+++ R/utils.R
@@ -13,9 +13,8 @@
 #' @param dates_to_parse Dates to parse
 #'
 #' @description
-#' This function is tailored for the date formatting used in this package,
-#' so it may fail if used with other datasets. See **Examples** for
-#' checking which formats would be considered.
+#' This function is tailored for date formats used in this package and may fail
+#' with other datasets. See **Examples** for formats that are supported.
 #'
 #' ## Date Formats
 #'